plthook: Wrap posix_spawn[p]()
The posix_spawn() basically does fork and exec at once.  And it requires
new environment variables to be used by the child.  In order to trace
the child, we need to pass uftrace-specific env to the child as well.
So wrap those functions and append the env at the end.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>